Move temporary function to migration.go
