+++ cmd/gardenlet/app/options.go
@@ -9,7 +9,6 @@ import (
 	"os"
 
 	"github.com/spf13/pflag"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/serializer"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
@@ -78,42 +77,4 @@ func (o *options) Validate() error {
 
 func (o *options) LogConfig() (string, string) {
 	return o.config.LogLevel, o.config.LogFormat
-}
-
-// syncBackupSecretRefAndCredentialsRef syncs the seed backup credentials when possible.
-// TODO(vpnachev): Remove this function after v1.121.0 has been released.
-func syncBackupSecretRefAndCredentialsRef(backup *gardencorev1beta1.SeedBackup) {
-	if backup == nil {
-		return
-	}
-
-	emptySecretRef := corev1.SecretReference{}
-
-	// secretRef is set and credentialsRef is not, sync both fields.
-	if backup.SecretRef != emptySecretRef && backup.CredentialsRef == nil {
-		backup.CredentialsRef = &corev1.ObjectReference{
-			APIVersion: "v1",
-			Kind:       "Secret",
-			Namespace:  backup.SecretRef.Namespace,
-			Name:       backup.SecretRef.Name,
-		}
-
-		return
-	}
-
-	// secretRef is unset and credentialsRef refer a secret, sync both fields.
-	if backup.SecretRef == emptySecretRef && backup.CredentialsRef != nil &&
-		backup.CredentialsRef.APIVersion == "v1" && backup.CredentialsRef.Kind == "Secret" {
-		backup.SecretRef = corev1.SecretReference{
-			Namespace: backup.CredentialsRef.Namespace,
-			Name:      backup.CredentialsRef.Name,
-		}
-
-		return
-	}
-
-	// in all other cases we can do nothing:
-	// - both fields are unset -> we have nothing to sync
-	// - both fields are set -> let the validation check if they are correct
-	// - credentialsRef refer to WorkloadIdentity -> secretRef should stay unset
 }
